requirements.txt
fastapi==0.135.2
uvicorn==0.42.0
langgraph==1.1.3
langchain-core==1.2.23
langchain-groq==1.1.2
groq==0.37.1
e2b-code-interpreter==2.6.0
python-dotenv==1.2.2
python-multipart==0.0.22